refactor(utils): extract system utilities to dedicated module
Move environment, file path, and process management functions to vllm/utils/system_utils.py:
- update_environment_variables: Update multiple environment variables
- set_env_var: Temporarily set environment variable (context manager)
- unique_filepath: Generate unique file paths
- set_process_title: Set process title with optional suffix
- decorate_logs: Add process name/PID prefix to stdout/stderr
- _add_prefix: Internal helper for log decoration

This reduces vllm/utils/__init__.py from 1310 to 1220 lines and
groups related system-level utilities in a single, cohesive module.

Contributes to #26900
