+++ tests/models/quantization/test_fp8.py
@@ -11,7 +11,6 @@ import pytest
 from tests.quantization.utils import is_quant_method_supported
 from vllm.platforms import current_platform
 from vllm.utils import STR_BACKEND_ENV_VAR
-
 from ..utils import check_logprobs_close
